// apps/mobile/api/generated/arenaAPI.schemas.ts
export interface TeamDto {
  /** Team ID */
  teamId?: string;
  /** Org ID */
  orgId?: string;
  /** 팀명 */
  name?: string;
  /** 멤버 수 */
  memberCount?: number;
}